+++ R/math.R
@@ -21,11 +21,19 @@ Math.units = function(x, ...) {
                "trunc" = , "round" = , "signif" = , "cumsum" = , 
                "cummax" = , "cummin" = TRUE, FALSE)
 
-  rad <- NULL # satisfy codetools warning
-  if (!OK && units(x) == units(set_units(1, rad))) {
+  rad = units(make_unit("rad"))
+  deg = units(make_unit("degree"))
+  if (!OK && (units(x) == rad || units(x) == deg)) {
     OK <- switch(.Generic, "sin" = , "cos" = , "tan" = TRUE, FALSE)
+    if (OK) {
+	  units(x) <- "rad" # convert deg -> rad
+	  x <- set_units(x) # result has unit 1
+	}
+  }
+  if (!OK && units(x) == unitless) {
+    OK <- switch(.Generic, "asin" = , "acos" = , "atan" = TRUE, FALSE)
     if (OK)
-	  x <- set_units(x)
+	  units(x) <- "rad" # unit of the answer (also unitless)
   }
 
   if (!OK) {